@@ -1,20 +1,24 @@
 package eu.siacs.conversations.xmpp.jingle;
 
 import android.content.Context;
+import android.media.ToneGenerator;
 import android.os.Build;
 import android.os.Handler;
 import android.os.Looper;
 import android.util.Log;
-import android.media.ToneGenerator;
 
 import com.google.common.base.Optional;
 import com.google.common.base.Preconditions;
+import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.util.concurrent.Futures;
 import com.google.common.util.concurrent.ListenableFuture;
 import com.google.common.util.concurrent.MoreExecutors;
 import com.google.common.util.concurrent.SettableFuture;
-import com.google.common.collect.ImmutableMap;
+
+import eu.siacs.conversations.Config;
+import eu.siacs.conversations.services.AppRTCAudioManager;
+import eu.siacs.conversations.services.XmppConnectionService;
 
 import org.webrtc.AudioSource;
 import org.webrtc.AudioTrack;
@@ -22,6 +26,7 @@ import org.webrtc.CandidatePairChangeEvent;
 import org.webrtc.DataChannel;
 import org.webrtc.DefaultVideoDecoderFactory;
 import org.webrtc.DefaultVideoEncoderFactory;
+import org.webrtc.DtmfSender;
 import org.webrtc.EglBase;
 import org.webrtc.IceCandidate;
 import org.webrtc.MediaConstraints;
@@ -35,24 +40,20 @@ import org.webrtc.SdpObserver;
 import org.webrtc.SessionDescription;
 import org.webrtc.VideoTrack;
 import org.webrtc.audio.JavaAudioDeviceModule;
-import org.webrtc.DtmfSender;
 
+import java.util.Collection;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Queue;
-import java.util.Set;
-import java.util.concurrent.Callable;
 import java.util.Map;
+import java.util.Set;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
+import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
-
-import eu.siacs.conversations.Config;
-import eu.siacs.conversations.services.AppRTCAudioManager;
-import eu.siacs.conversations.services.XmppConnectionService;
 
 @SuppressWarnings("UnstableApiUsage")
 public class WebRTCWrapper {
@@ -62,26 +63,6 @@ public class WebRTCWrapper {
     private final ExecutorService executorService = Executors.newSingleThreadExecutor();
     private final ExecutorService localDescriptionExecutorService =
             Executors.newSingleThreadExecutor();
-
-    private static final Set<String> HARDWARE_AEC_BLACKLIST =
-            new ImmutableSet.Builder<String>()
-                    .add("Pixel")
-                    .add("Pixel XL")
-                    .add("Moto G5")
-                    .add("Moto G (5S) Plus")
-                    .add("Moto G4")
-                    .add("TA-1053")
-                    .add("Mi A1")
-                    .add("Mi A2")
-                    .add("E5823") // Sony z5 compact
-                    .add("Redmi Note 5")
-                    .add("FP2") // Fairphone FP2
-                    .add("FP4") //Fairphone FP4
-                    .add("MI 5")
-                    .add("GT-I9515") // Samsung Galaxy S4 Value Edition (jfvelte)
-                    .add("GT-I9515L") // Samsung Galaxy S4 Value Edition (jfvelte)
-                    .add("GT-I9505") // Samsung Galaxy S4 (jfltexx)
-                    .build();
 
     private static final int TONE_DURATION = 500;
     private static final Map<String,Integer> TONE_CODES;
@@ -101,6 +82,26 @@ public class WebRTCWrapper {
         builder.put("#", ToneGenerator.TONE_DTMF_P);
         TONE_CODES = builder.build();
     }
+
+    private static final Set<String> HARDWARE_AEC_BLACKLIST =
+            new ImmutableSet.Builder<String>()
+                    .add("Pixel")
+                    .add("Pixel XL")
+                    .add("Moto G5")
+                    .add("Moto G (5S) Plus")
+                    .add("Moto G4")
+                    .add("TA-1053")
+                    .add("Mi A1")
+                    .add("Mi A2")
+                    .add("E5823") // Sony z5 compact
+                    .add("Redmi Note 5")
+                    .add("FP2") // Fairphone FP2
+                    .add("FP4") // Fairphone FP4
+                    .add("MI 5")
+                    .add("GT-I9515") // Samsung Galaxy S4 Value Edition (jfvelte)
+                    .add("GT-I9515L") // Samsung Galaxy S4 Value Edition (jfvelte)
+                    .add("GT-I9505") // Samsung Galaxy S4 (jfltexx)
+                    .build();
 
     private final EventCallback eventCallback;
     private final AtomicBoolean readyToReceivedIceCandidates = new AtomicBoolean(false);
@@ -213,7 +214,6 @@ public class WebRTCWrapper {
                                     + ")");
                     if (track instanceof VideoTrack) {
                         remoteVideoTrack = (VideoTrack) track;
-                        eventCallback.onTrackModification();
                     }
                 }
 
@@ -263,8 +263,7 @@ public class WebRTCWrapper {
             PeerConnectionFactory.initialize(
                     PeerConnectionFactory.InitializationOptions.builder(service)
                             .setFieldTrials("WebRTC-BindUsingInterfaceName/Enabled/")
-                            .createInitializationOptions()
-            );
+                            .createInitializationOptions());
         } catch (final UnsatisfiedLinkError e) {
             throw new InitializationException("Unable to initialize PeerConnectionFactory", e);
         }
@@ -370,8 +369,6 @@ public class WebRTCWrapper {
         }
     }
 
-
-
     private boolean addAudioTrack(final PeerConnection peerConnection) {
         final AudioSource audioSource =
                 requirePeerConnectionFactory().createAudioSource(new MediaConstraints());
@@ -407,10 +404,7 @@ public class WebRTCWrapper {
                         .createVideoTrack(
                                 TrackWrapper.id(VideoTrack.class),
                                 videoSourceWrapper.getVideoSource());
-        // TODO do we want to create Transceiver manually and be able to set direction and keep a
-        // reference to it for later removal
         this.localVideoTrack = TrackWrapper.addTrack(peerConnection, videoTrack);
-        this.eventCallback.onTrackModification();
         return true;
     }
 
@@ -434,8 +428,6 @@ public class WebRTCWrapper {
             }
         }
     }
-
-
 
     private static PeerConnection.RTCConfiguration buildConfiguration(
             final List<PeerConnection.IceServer> iceServers, final boolean trickle) {
@@ -640,7 +632,7 @@ public class WebRTCWrapper {
                                 public void onSetSuccess() {
                                     final var delay =
                                             waitForCandidates
-                                                    ? iceGatheringComplete
+                                                    ? Futures.catching(Futures.withTimeout(iceGatheringComplete, 2, TimeUnit.SECONDS, JingleConnectionManager.SCHEDULED_EXECUTOR_SERVICE), Exception.class, (Exception e) -> { return null; }, MoreExecutors.directExecutor())
                                                     : Futures.immediateVoidFuture();
                                     final var delayedSessionDescription =
                                             Futures.transformAsync(
@@ -671,39 +663,6 @@ public class WebRTCWrapper {
                     return description;
                 },
                 localDescriptionExecutorService);
-    }
-
-
-    synchronized ListenableFuture<SessionDescription> rollback() {
-        return Futures.transformAsync(
-                getPeerConnectionFuture(),
-                peerConnection -> {
-                    final SettableFuture<SessionDescription> future = SettableFuture.create();
-                    if (peerConnection == null) {
-                        return Futures.immediateFailedFuture(
-                                new IllegalStateException("PeerConnection was null"));
-                    }
-                    peerConnection.setLocalDescription(
-                            new SetSdpObserver() {
-                                @Override
-                                public void onSetSuccess() {
-                                    final SessionDescription description =
-                                            peerConnection.getLocalDescription();
-                                    Log.d(EXTENDED_LOGGING_TAG, "rollback to local description:");
-                                    logDescription(description);
-                                    future.set(description);
-                                }
-
-                                @Override
-                                public void onSetFailure(final String message) {
-                                    future.setException(
-                                            new FailureToSetDescriptionException(message));
-                                }
-                            },
-                            new SessionDescription(SessionDescription.Type.ROLLBACK, ""));
-                    return future;
-                },
-                MoreExecutors.directExecutor());
     }
 
     public static void logDescription(final SessionDescription sessionDescription) {
@@ -764,15 +723,6 @@ public class WebRTCWrapper {
         return peerConnection;
     }
 
-    @Nonnull
-    private PeerConnectionFactory requirePeerConnectionFactory() {
-        final PeerConnectionFactory peerConnectionFactory = this.peerConnectionFactory;
-        if (peerConnectionFactory == null) {
-            throw new IllegalStateException("Make sure PeerConnectionFactory is initialized");
-        }
-        return peerConnectionFactory;
-    }
-
     public boolean applyDtmfTone(String tone) {
         if (toneManager == null || peerConnection == null || localAudioTrack == null) {
             return false;
@@ -780,6 +730,15 @@ public class WebRTCWrapper {
         localAudioTrack.rtpSender.dtmf().insertDtmf(tone, TONE_DURATION, 100);
         toneManager.startTone(TONE_CODES.get(tone), TONE_DURATION);
         return true;
+    }
+
+    @Nonnull
+    private PeerConnectionFactory requirePeerConnectionFactory() {
+        final PeerConnectionFactory peerConnectionFactory = this.peerConnectionFactory;
+        if (peerConnectionFactory == null) {
+            throw new IllegalStateException("Make sure PeerConnectionFactory is initialized");
+        }
+        return peerConnectionFactory;
     }
 
     void addIceCandidate(IceCandidate iceCandidate) {
@@ -830,7 +789,6 @@ public class WebRTCWrapper {
         mainHandler.post(() -> appRTCAudioManager.switchSpeakerPhonePreference(preference));
     }
 
-
     public interface EventCallback {
         void onIceCandidate(IceCandidate iceCandidate);
 
@@ -841,8 +799,6 @@ public class WebRTCWrapper {
                 Set<AppRTCAudioManager.AudioDevice> availableAudioDevices);
 
         void onRenegotiationNeeded();
-
-        void onTrackModification();
     }
 
     private abstract static class SetSdpObserver implements SdpObserver {
