@@ -70,6 +70,9 @@ public class IceUdpTransportInfo extends GenericTransportInfo {
         for (final String iceOption : IceOption.of(media)) {
             iceUdpTransportInfo.addChild(new IceOption(iceOption));
         }
+        for (final String candidate : media.attributes.get("candidate")) {
+            iceUdpTransportInfo.addChild(Candidate.fromSdpAttributeValue(candidate, ufrag));
+        }
         return iceUdpTransportInfo;
     }
 
@@ -96,7 +99,7 @@ public class IceUdpTransportInfo extends GenericTransportInfo {
 
     public List<String> getIceOptions() {
         final ImmutableList.Builder<String> optionBuilder = new ImmutableList.Builder<>();
-        for (final Element child : this.children) {
+        for (final Element child : getChildren()) {
             if (Namespace.JINGLE_TRANSPORT_ICE_OPTION.equals(child.getNamespace())
                     && IceOption.WELL_KNOWN.contains(child.getName())) {
                 optionBuilder.add(child.getName());
@@ -114,7 +117,7 @@ public class IceUdpTransportInfo extends GenericTransportInfo {
     public boolean isStub() {
         return Strings.isNullOrEmpty(this.getAttribute("ufrag"))
                 && Strings.isNullOrEmpty(this.getAttribute("pwd"))
-                && this.children.isEmpty();
+                && getChildren().isEmpty();
     }
 
     public List<Candidate> getCandidates() {
